Refactored initial conditions to own module.
